refactor(tool-settings): extract builtin presets and tip generators (#453)
Move 7 tip generators, 5 texture generators, BUILTIN_TEXTURES,
BUILTIN_PRESETS, createPresetId, and abrBrushToPreset out of
tool-settings-store.ts into src/tools/brush/builtin-presets.ts.
Store file drops from 955 to 564 lines (41% reduction). Re-exports
preserve the existing public API so no consumer changes needed.
